requirements.txt
alembic==1.16.4
fastapi[standard]==0.116.1
uvicorn==0.35.0
psycopg-binary==3.2.9
pydantic==2.11.7
sqlmodel==0.0.24
python-dotenv==1.1.1
pydantic-settings==2.10.1
